LayoutTests/sputnik/Conformance/12_Statement/12.14_The_try_Statement/S12.14_A16_T6-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Unexpected token 'catch'
S12.14_A16_T6

PASS Expected parsing failure

TEST COMPLETE
